Sharing structure to own files
