+++ src/index.ts
@@ -1,10 +1,11 @@
 'use strict';
 
 import {
-  CertificateInterface,
+  CertificateInterface, CreateCertificateInterface, CreatedCertificateInterface,
   GetCertificateInterface, XTInterface
 } from './interfaces';
-import {Base64DecodeStr, Base64EncodeStr, CreateCertificate} from './utils';
+import {Base64DecodeStr, Base64EncodeStr} from './utils';
+import {CreateCertificate} from './create_certificate/CreateCertificate';
 import {CertApplicationRequest} from './get_certificate/certApplicationRequest';
 import {CertRequestEnvelope} from './get_certificate/certRequestEnvelope';
 import {CertApplicationResponse} from './get_certificate/certApplicationResponse';
@@ -16,8 +17,14 @@ import * as https from 'https';
 import axios from 'axios';
 
 
-async function CreateOwnCertificate(): Promise<any> {
-  await CreateCertificate();
+/**
+ * Create your own client side certificate
+ * @param cc, certificate parameters
+ * @constructor
+ */
+async function CreateOwnCertificate(cc: CreateCertificateInterface): Promise<CreatedCertificateInterface> {
+  const ownCertificate = new CreateCertificate(cc);
+  return await ownCertificate.createCertificate();
 }
